# apps/server/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects/{id}/actions/evaluate-financial-completion": {
      "post": {
        "operationId": "evaluateProjectFinancialCompletion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialCompletionDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/jobs/{id}/actions/evaluate-financial-completion": {
      "post": {
        "operationId": "evaluateJobFinancialCompletion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialCompletionDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
...
  },
  "info": {
    "title": "Lincoln Dirt and Gravel API",
    "description": "Lincoln Dirt and Gravel operating system API",
    "version": "1.7.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "JobFinancialCompletionDto": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "jobNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "financiallyComplete": {
            "type": "boolean"
          },
          "postedObligationCount": {
            "type": "number"
          },
          "outstandingCents": {
            "type": "number"
          },
          "blockers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "jobId",
          "jobNumber",
          "status",
          "financiallyComplete",
          "postedObligationCount",
          "outstandingCents",
          "blockers"
        ]
      },
      "FinancialCompletionDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "projectNumber": {
            "type": "string"
          },
          "projectStatus": {
            "type": "string"
          },
          "financiallyComplete": {
            "type": "boolean"
          },
          "outstandingInvoiceCents": {
            "type": "number"
          },
          "unresolvedPaymentCents": {
            "type": "number"
          },
          "unresolvedDepositCents": {
            "type": "number"
          },
          "unresolvedCustomerCreditCents": {
            "type": "number"
          },
          "activeRefundCount": {
            "type": "number"
          },
          "blockers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobFinancialCompletionDto"
            }
          },
          "evaluatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "projectId",
          "projectNumber",
          "projectStatus",
          "financiallyComplete",
          "outstandingInvoiceCents",
          "unresolvedPaymentCents",
          "unresolvedDepositCents",
          "unresolvedCustomerCreditCents",
          "activeRefundCount",
          "blockers",
          "jobs",
          "evaluatedAt"
        ]
      },
...
    }
  }
}